+++ lib/mocktail/value/explanation_data.rb
@@ -1,0 +1,4 @@
+module Mocktail
+  module ExplanationData
+  end
+end