+++ lib/mocktail/value/fake_method_data.rb
@@ -6,5 +6,7 @@ module Mocktail
     :calls,
     :stubbings,
     keyword_init: true
-  )
+  ) do
+    include ExplanationData
+  end
 end